back/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
importlib_metadata==8.5.0
Markdown==3.6
mysqlclient==2.2.7
pillow==11.2.1
pip==24.3.1
sqlparse==0.5.2
typing_extensions==4.12.2
zipp==3.21.0
